// src/app/core/api/model/baseConfigDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CarEngineDto } from './carEngineDto';
import { SpecialEquipmentDto } from './specialEquipmentDto';
import { CarRimDto } from './carRimDto';
import { CarColorDto } from './carColorDto';


export interface BaseConfigDto { 
    carColors?: Array<CarColorDto>;
    carEngines?: Array<CarEngineDto>;
    carRims?: Array<CarRimDto>;
    specialEquipment?: Array<SpecialEquipmentDto>;
}